+++ tests/test_generic.py
@@ -86,7 +86,7 @@ class TestFrequencyAnalysis:
         )
 
 
-class TestQuantile:
+class TestParametricQuantile:
     def test_synth(self):
         mu = 23
         sigma = 2